src/auth/requirements.txt
grpcio 
grpcio-tools
python-dotenv
sqlmodel
psycopg2
pydantic-extra-types
pydantic[email]
phonenumbers
sqlalchemy
sqlalchemy-utils
passlib
pyjwt